Move modal to shared.
