+++ packages/front-end/src/components/optionsTrading/Modals/Shared/Modal/Modal.tsx
